登録・ログインをダイアログ画面で行うように修正
